pyproject.toml
[tool.pdm]
distribution = true

[tool.black]
line-length = 120
target-version = ["py313"]
skip-string-normalization = false
exclude = '''
/(
  \.git
  | \.hg
  | \.venv
  | venv
  | build
  | dist
  | src/apigateway/stubs
)/
'''

[tool.isort]
profile = "black"
line_length = 120
known_first_party = ["app"]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip = [".venv", "src/apigateway/stubs"]

[tool.flake8]
max-line-length = 120
extend-ignore = ["E203", "W503", "E501"]
per-file-ignores = [
    "__init__.py:F401",
    "tests/*:T20,ARG001",
    "src/apigateway/stubs/*:E501,ARG001,ARG002"
]
exclude = [
  ".venv",
  "__pycache__",
  "build",
  "dist",
  "src/apigateway/stubs",
]

[tool.ruff]
line-length = 120
target-version = "py313"
preview = true
exclude = [".venv", "build", "dist", "src/apigateway/stubs"]

[tool.ruff.lint]
select = [
    "E",
    "F",
    "B",
    "UP",
    "SIM",
    "T20",
    "ARG",
    "C4",
    "RUF"
]
ignore = [
    "E203",
    "B008",
    "SIM105",
    "RUF001",
    "RUF100",
    "RUF003",
    "RUF002"
]
extend-ignore = [
    "ARG002"
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["T20", "ARG001"]
"src/apigateway/resolvers/*" = ["ARG001", "ARG002"]
"src/apigateway/stubs/*" = ["E501","ARG001", "ARG002"]

[tool.mypy]
python_version = "3.13"
strict = true
warn_unused_configs = true
warn_return_any = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
show_error_codes = true
explicit_package_bases = true
mypy_path = ["src"]
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = ["apigateway.stubs.*"]
ignore_errors = true

[[tool.mypy.overrides]]
module = ["tests.*"]
ignore_errors = true

[tool.pytest.ini_options]
addopts = "--import-mode=importlib"
asyncio_mode = "strict"
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.coverage.run]
source = ["src/apigateway"]
relative_files = true

[tool.coverage.report]
include = [
    "src/apigateway/*",
]

[project]
name = "api-gateway"
version = "0.0.1"
description = "api-gateway for esclient"
authors = [
    {name = "esclient"},
]
dependencies = [
    "ariadne<1.0.0,>=0.26.2",
    "uvicorn<1.0.0,>=0.35.0",
    "grpcio-tools>=1.74.0",
    "pydantic>=2.11.9",
    "pydantic-settings>=2.10.1",
    "tenacity>=9.1.2",
]
requires-python = "<4.0,>=3.13"
readme = "README.md"
license = {text = "MIT"}

[dependency-groups]
lint = [
    "flake8>=7.3.0",
]
dev = [
    "black>=25.9.0",
    "faker>=25.0.0",
    "flake8>=7.3.0",
    "flake8-pyproject>=1.2.3",
    "isort>=6.0.1",
    "mypy>=1.18.2",
    "pytest>=8.3.0",
    "pytest-cov==5.0.0",
    "pytest-asyncio>=0.25.0",
    "pytest-faker>=2.0.0",
    "ruff>=0.13.1",
    "types-grpcio>=1.0.0.20250914",
    "types-protobuf>=6.32.1.20250918",
]
